third_party/blink/web_tests/virtual/pna-iframes-warning/external/wpt/fetch/private-network-access/iframe.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS local to local: no preflight required.
PASS local to private: no preflight required.
PASS local to public: no preflight required.
FAIL private to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL private to local: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL private to local: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS private to local: success.
PASS private to private: no preflight required.
PASS private to public: no preflight required.
FAIL public to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL public to local: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL public to local: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS public to local: success.
FAIL public to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL public to private: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL public to private: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS public to private: success.
PASS public to public: no preflight required.
FAIL treat-as-public-address to local: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL treat-as-public-address to local: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL treat-as-public-address to local: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS treat-as-public-address to local: success.
PASS treat-as-public-address to local (same-origin): no preflight required.
FAIL treat-as-public-address to private: failed preflight. assert_equals: expected "timeout" but got "loaded"
FAIL treat-as-public-address to private: missing CORS headers. assert_equals: expected "timeout" but got "loaded"
FAIL treat-as-public-address to private: missing PNA header. assert_equals: expected "timeout" but got "loaded"
PASS treat-as-public-address to private: success.
PASS treat-as-public-address to public: no preflight required.
PASS treat-as-public-address to local: optional preflight
PASS local to local, grandparent navigates: no preflight required.
PASS local to local (same-origin), grandparent navigates: no preflight required.
FAIL public to local, grandparent navigates: failure. assert_equals: expected "timeout" but got "loaded"
PASS public to local, grandparent navigates: success.
Harness: the test ran to completion.
